+++ RELEASE_NOTES.md
@@ -31,6 +31,10 @@ RELEASE NOTES
 Logtalk compiler and runtime
 ----------------------------
 
+* NEW: Updated the compiler to perform error and portability checking
+for the `set_prolog_flag/2` and `current_prolog_flag/2` standard Prolog
+built-in predicates.
+
 * CHANGED: Compiler messages about redefined entities that result from
 reloading the same source file that contained the previous definitions
 are no longer warning but informational messages.